requirements.txt
discord.py==2.0.0
dynaconf==3.1.9
